requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.7.14
chardet==3.0.4
charset-normalizer==3.4.2
click==8.2.1
deep-translator==1.11.4
fastapi==0.116.1
greenlet>=3.1.1
h11==0.16.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hyperframe==5.2.0
idna==2.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.4
rfc3986==1.5.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
scrapy==2.11.2
scrapy-playwright==0.0.34
playwright==1.48.0
